kup-list: moved utlity methods from kup-list-declarations.ts to kup-list-helper.ts.
